Completion Hanlder Threading Fixes
This fixes some issues where completion handler was calling back to the main thread even when the engine was rendering to a file. This causes the completion handler to not run until after the rendering to a file is finished, which means looping players don't work. This should fix that.